Fixed the bug that occurred with multiple calls to startPlayer() (iOS)
Before this commit, when the startPlayer() method was called (in Dart) many times without waiting for each async call to finish loading (and then, for the tracks to be fully retrieved), the player created multiple download tasks, one for each given track. Then, as the tasks were completed, the player started to play the just downloaded track.
After this commit, whenever the startPlayer() method is called, the download task created for the previously provided track is canceled. So, reproducing the same context that I explained above, only the track provided by the last call will be played.
This commit is only refers to iOS.